IMTA-15450-upgrade-java-17

See merge request imports/spring-boot-common!66

+++ health/src/main/java/uk/gov/defra/tracesx/common/health/HealthCheckConfiguration.java
@@ -30,8 +30,8 @@ public class HealthCheckConfiguration {
   @Qualifier("defaultHealthCheckRestTemplate")
   public RestTemplate defaultHealthCheckRestTemplate() {
     return new RestTemplateBuilder()
-        .setConnectTimeout(Duration.ofSeconds((long) connectTimeout))
-        .setReadTimeout(Duration.ofSeconds((long) readTimeout))
+        .setConnectTimeout(Duration.ofSeconds(connectTimeout))
+        .setReadTimeout(Duration.ofSeconds(readTimeout))
         .build();
   }
